src/requirements.txt
LyricsGenius
python-dotenv
pika